All Managers Except UserManager

+++ QPEKA_UI/src/com/qpeka/db/Files.java
@@ -31,7 +31,7 @@ public class Files implements Serializable {
 	private String filemime;
 	private int filesize;
 	private int status;
-	private int timestamp;
+	private long timestamp;
 
 	/**
 	 * These attributes represents whether the above attributes has been
@@ -98,7 +98,7 @@ public class Files implements Serializable {
 		return userid;
 	}
 
-	public void setUserid(int userid) {
+	public void setUserid(long userid) {
 		this.userid = userid;
 		this.useridModified = true;
 	}
@@ -157,11 +157,11 @@ public class Files implements Serializable {
 		this.statusModified = true;
 	}
 
-	public int getTimestamp() {
+	public long getTimestamp() {
 		return timestamp;
 	}
 
-	public void setTimestamp(int timestamp) {
+	public void setTimestamp(long timestamp) {
 		this.timestamp = timestamp;
 		this.timestampModified = true;
 	}
@@ -371,11 +371,11 @@ public class Files implements Serializable {
 		}
 
 		_hashCode = 29 * _hashCode + (filemimeModified ? 1 : 0);
-		_hashCode = 29 * _hashCode + filesize;
+		_hashCode = 29 * _hashCode + filesize; 
 		_hashCode = 29 * _hashCode + (filesizeModified ? 1 : 0);
 		_hashCode = 29 * _hashCode + status;
 		_hashCode = 29 * _hashCode + (statusModified ? 1 : 0);
-		_hashCode = 29 * _hashCode + timestamp;
+		_hashCode = 29 * _hashCode + (int)timestamp;
 		_hashCode = 29 * _hashCode + (timestampModified ? 1 : 0);
 		return _hashCode;
 	}
